Unificação do projeto

+++ src/main/java/br/com/tiacademy/exameagenda/core/crud/CrudService.java
@@ -42,6 +42,10 @@ public abstract class CrudService<T, ID> {
 	}
 
 	public void excluir(ID id) {
+		T recuperado = porId(id);
+		if (Objects.isNull(recuperado)) {
+			throw new RuntimeException("Objeto não foi encontrado!");
+		}
 		repository.deleteById(id);
 	}
